Refactor the upsert schema
