+++ apps/auth/src/features/pg-schema/down.js
@@ -1,0 +1,3 @@
+module.exports = async (pg) => {
+  await pg.query('DROP SCHEMA IF EXISTS "public" CASCADE');
+};
