+++ apps/auth/src/index.js
@@ -1,18 +1,35 @@
+const envalid = require("envalid");
 const forrest = require("@forrestjs/core");
+
+// Library Services
 const serviceFastify = require("@forrestjs/service-fastify");
 const serviceHealthz = require("@forrestjs/service-fastify-healthz");
 const serviceJwt = require("@forrestjs/service-jwt");
 const serviceCookie = require("@forrestjs/service-fastify-cookie");
 
+// Local Services
 const servicePg = require("./service-pg");
 
-const upsertSchema = require("./upsert-schema");
+// Local Features
+const pgSchema = require("./features/pg-schema");
 const publicPages = require("./features/public-pages");
 const privatePages = require("./features/private-pages");
 const tokenApi = require("./features/token-api");
 
+// Validate the environment
+const env = envalid.cleanEnv(process.env, {
+  JWT_SECRET: envalid.str(),
+  JWT_DURATION: envalid.str()
+});
+
 forrest.run({
   // trace: "compact",
+  settings: {
+    jwt: {
+      secret: env.JWT_SECRET,
+      duration: env.JWT_DURATION
+    }
+  },
   services: [
     serviceFastify,
     serviceHealthz,
@@ -20,5 +37,5 @@ forrest.run({
     servicePg,
     serviceCookie
   ],
-  features: [upsertSchema, publicPages, privatePages, tokenApi]
+  features: [pgSchema, publicPages, privatePages, tokenApi]
 });
